+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/NotificationHook/EmailNotificationHook.cs
@@ -25,7 +25,7 @@ namespace Azure.AI.MetricsAdvisor.Administration
             EmailsToAlert = new ChangeTrackingList<string>();
         }
 
-        internal EmailNotificationHook(HookType hookType, string id, string name, string description, string externalLink, IReadOnlyList<string> administrators, EmailHookParameter hookParameter)
+        internal EmailNotificationHook(HookType hookType, string id, string name, string description, string externalLink, IList<string> administrators, EmailHookParameter hookParameter)
             : base(hookType, id, name, description, externalLink, administrators)
         {
             HookType = hookType;