+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/NotificationHook/WebNotificationHook.cs
@@ -30,7 +30,7 @@ namespace Azure.AI.MetricsAdvisor.Administration
             Headers = new ChangeTrackingDictionary<string, string>();
         }
 
-        internal WebNotificationHook(HookType hookType, string id, string name, string description, string externalLink, IReadOnlyList<string> administrators, WebhookHookParameter hookParameter)
+        internal WebNotificationHook(HookType hookType, string id, string name, string description, string externalLink, IList<string> administrators, WebhookHookParameter hookParameter)
             : base(hookType, id, name, description, externalLink, administrators)
         {
             HookType = hookType;