+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/tests/Samples/Sample04_HookCrudOperations.cs
@@ -67,8 +67,8 @@ namespace Azure.AI.MetricsAdvisor.Samples
             Console.WriteLine($"Hook name: {hook.Name}");
             Console.WriteLine($"Hook description: {hook.Description}");
 
-            Console.WriteLine($"Hook administrators emails:");
-            foreach (string admin in hook.AdministratorEmails)
+            Console.WriteLine($"Hook administrators:");
+            foreach (string admin in hook.Administrators)
             {
                 Console.WriteLine($" - {admin}");
             }